+++ NEWS.md
@@ -1,9 +1,11 @@
-# pomdp 0.99.0-1 (xx/xx/2020)
+# pomdp 0.99.1 (05/11/2021)
+
+## New Features
+* Added a wrapper for the sarsop library.
 
 ## Changes
 * Improved error messages when accessing fields not parsed by read_POMDP.
 * policy() no longer returns the graph, but just alphas and the optimal action.
-* Added a wrapper for the sarsop library.
 
 ## Bugfix
 * Resolved issues with factors for R 4.0. We now mostly use character instead of factors.